PE Platform cleanup for Linux build warnings
 * Support -Wall -Werror -Wno-unknown-pragmas
 * Partial support for -Wextra, investigating full support
 * XL, GCC compilers are error/warning free
 * Intel compilers are error/warning free with some exceptions
   and some warnings disabled
 * BGQ, SLES11, RedHat, AIX builds tested
 * x86(64-bit), Power(32,64-bit), A2(64-bit) architectures built

+++ algorithms/composite/MultiColorCompositeT.h
@@ -135,7 +135,7 @@ namespace CCMI
                        char                                  * dst)
       {
         TRACE_FN_ENTER();
-        TRACE_FORMAT("<%p> root %u, bytes %lu", this,root, stypecount * stype->GetDataSize());
+        TRACE_FORMAT("<%p> root %u, bytes %lu", this,root, (unsigned long)stypecount * stype->GetDataSize());
         pwcfn (topology, stypecount * stype->GetDataSize(), _colors, _numColors);
         //printf ("Using %d colors, %d\n", _numColors, _colors[0]);
         if (_numColors > NUMCOLORS)
@@ -205,7 +205,7 @@ namespace CCMI
                   char                                  * dst)
       {
         TRACE_FN_ENTER();
-        TRACE_FORMAT("<%p> root %u, bytes %lu", this,root, stypecount * stype->GetDataSize());
+        TRACE_FORMAT("<%p> root %u, bytes %lu", this,root, (unsigned long)stypecount * stype->GetDataSize());
         _doneCount = 0;
 	
 	if (_bytes != (stypecount * stype->GetDataSize()) || _root != root) {
